requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.0
Django==3.0.8
-e git://github.com/DavidJayBrady/django-email-verification@1563b7f368db9f4c69b79cc0e0c8e34967a84541#egg=django_email_verification
django-nose==1.4.6
django-password-reset==2.0
djangorestframework==3.11.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
ndg-httpsclient==0.5.1
nose==1.3.7
openapi-codec==1.3.2
Pillow==7.2.0
pyasn1==0.4.8
pycparser==2.20
pyOpenSSL==19.1.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
